requirements.txt
asgiref==3.7.2
contourpy==1.1.1
cycler==0.12.1
Django==3.2.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
fonttools==4.45.1
gunicorn==21.2.0
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.7.4
networkx==3.1
numpy==1.24.4
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
scikit-fuzzy==0.4.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
zipp==3.17.0